Crate wasmtime_wasi

Wasmtime’s WASI Implementation

This crate provides a Wasmtime host implementation of WASI 0.2 (aka Preview 2), and a compatibility shim that provides an implementation of WASI 0.1 (aka Preview 1).

Modules

bindings

command

pipe

Virtual pipes.

preview0

preview1

runtime

This module provides an “ambient Tokio runtime” with_ambient_tokio_runtime. Embedders of wasmtime-wasi may do so from synchronous Rust, and not use tokio directly. The implementation of wasmtime-wasi requires a tokio executor in a way that is deeply tied to its design. When used from a sychrnonous wasmtime context, this module provides the wrapper function in_tokio used throughout the shim implementations of synchronous component binding Host traits in terms of the async ones.

Structs

AsyncStdinStream

An impl of StdinStream built on top of crate::pipe::AsyncReadStream.

AsyncStdoutStream

A wrapper of crate::pipe::AsyncWriteStream that implements StdoutStream. Note that the HostOutputStream impl for this is not correct when used for interleaved async IO.

Deterministic

Implement insecure-random using a deterministic cycle of bytes.

DirPerms

FilePerms

I32Exit

An error returned from the proc_exit host syscall.

Network

Pollable

A host representation of the wasi:io/poll.pollable resource.

ResourceTable

The ResourceTable type maps a Resource<T> to its T.

Stderr

This implementation will yield output streams that block on writes, as they inherit the implementation directly from the rust std library. A different implementation of StdoutStream will be necessary if truly async output streams are required.

Stdin

Only public interface is the HostInputStream impl.

Stdout

This implementation will yield output streams that block on writes, as they inherit the implementation directly from the rust std library. A different implementation of StdoutStream will be necessary if truly async output streams are required.

TrappableError

A helper error type used by many other modules through type aliases.

WasiCtx

WasiCtxBuilder

WasiP1Ctx

Enums

InputStream

IsATTY

ResourceTableError

Errors returned by operations on ResourceTable

StreamError

SystemTimeSpec

A value for specifying a time.

Traits

HostInputStream

Host trait for implementing the wasi:io/streams.input-stream resource: A bytestream which can be read from.

HostMonotonicClock

HostOutputStream

Host trait for implementing the wasi:io/streams.output-stream resource: A bytestream which can be written to.

HostWallClock

RngCore

The core of a random number generator.

StdinStream

A trait used to represent the standard input to a guest program.

StdoutStream

Similar to StdinStream, except for output.

Subscribe

WasiView

Functions

stderr

stdin

stdout

subscribe

Creates a pollable resource which is susbcribed to the provided resource.

thread_rng

Type Aliases

ClosureFuture

FsError

FsResult

MakeFuture

OutputStream

PollableFuture

SocketError

SocketResult

StreamResult