Type Alias wasmtime_wasi::bindings::io::error::Error

pub type Error = Error;

A resource which represents some error information.

The only method provided by this resource is to-debug-string, which provides some human-readable information about the error.

In the wasi:io package, this resource is returned through the wasi:io/streams/stream-error type.

To provide more specific error information, other interfaces may provide functions to further “downcast” this error into more specific error information. For example, errors returned in streams derived from filesystem types to be described using the filesystem’s own error-code type, using the function wasi:filesystem/types/filesystem-error-code, which takes a parameter borrow<error> and returns option<wasi:filesystem/types/error-code>.

The set of functions which can “downcast” an error into a more concrete type is open. Representation of the error resource type in the wasi:io/error interface.

This is currently anyhow::Error to retain full type information for errors.

Aliased Type

struct Error { /* private fields */ }