Type Alias wasmtime_wasi::bindings::sockets::ip_name_lookup::ErrorCode

pub type ErrorCode = ErrorCode;

Aliased Type

enum ErrorCode {

    Unknown,
    AccessDenied,
    NotSupported,
    InvalidArgument,
    OutOfMemory,
    Timeout,
    ConcurrencyConflict,
    NotInProgress,
    WouldBlock,
    InvalidState,
    NewSocketLimit,
    AddressNotBindable,
    AddressInUse,
    RemoteUnreachable,
    ConnectionRefused,
    ConnectionReset,
    ConnectionAborted,
    DatagramTooLarge,
    NameUnresolvable,
    TemporaryResolverFailure,
    PermanentResolverFailure,

}

Variants

Unknown

Unknown error

AccessDenied

Access denied.

POSIX equivalent: EACCES, EPERM

NotSupported

The operation is not supported.

POSIX equivalent: EOPNOTSUPP

InvalidArgument

One of the arguments is invalid.

POSIX equivalent: EINVAL

OutOfMemory

Not enough memory to complete the operation.

POSIX equivalent: ENOMEM, ENOBUFS, EAI_MEMORY

Timeout

The operation timed out before it could finish completely.

ConcurrencyConflict

This operation is incompatible with another asynchronous operation that is already in progress.

POSIX equivalent: EALREADY

NotInProgress

Trying to finish an asynchronous operation that:

has not been started yet, or:

was already finished by a previous finish-* call.

Note: this is scheduled to be removed when futures are natively supported.

WouldBlock

The operation has been aborted because it could not be completed immediately.

Note: this is scheduled to be removed when futures are natively supported.

InvalidState

The operation is not valid in the socket’s current state.

NewSocketLimit

A new socket resource could not be created because of a system limit.

AddressNotBindable

A bind operation failed because the provided address is not an address that the network can bind to.

AddressInUse

A bind operation failed because the provided address is already in use or because there are no ephemeral ports available.

RemoteUnreachable

The remote address is not reachable

ConnectionRefused

The TCP connection was forcefully rejected

ConnectionReset

The TCP connection was reset.

ConnectionAborted

A TCP connection was aborted.

DatagramTooLarge

The size of a datagram sent to a UDP socket exceeded the maximum supported size.

NameUnresolvable

Name does not exist or has no suitable associated IP addresses.

TemporaryResolverFailure

A temporary failure in name resolution occurred.

PermanentResolverFailure

A permanent failure in name resolution occurred.