Struct wasmtime_wasi::preview1::WasiP1Ctx

pub struct WasiP1Ctx { /* private fields */ }

Available on crate feature preview1 only.

Structure containing state for WASIp1.

This structure is created through WasiCtxBuilder::build_p1 and is configured through the various methods of WasiCtxBuilder. This structure itself implements generated traits for WASIp1 as well as WasiView to have access to WASIp2.

Instances of WasiP1Ctx are typically stored within the T of Store<T>.

Examples

use wasmtime::{Result, Linker};
use wasmtime_wasi::preview1::{self, WasiP1Ctx};
use wasmtime_wasi::WasiCtxBuilder;

struct MyState {
    // ... custom state as necessary ...

    wasi: WasiP1Ctx,
}

impl MyState {
    fn new() -> MyState {
        MyState {
            // .. initialize custom state if needed ..

            wasi: WasiCtxBuilder::new()
                .arg("./foo.wasm")
                // .. more customization if necesssary ..
                .build_p1(),
        }
    }
}

fn add_to_linker(linker: &mut Linker<MyState>) -> Result<()> {
    preview1::add_to_linker_sync(linker, |my_state| &mut my_state.wasi)?;
    Ok(())
}