Trait wasmtime_wasi::preview2::bindings::io::streams::Host

pub trait Host {

    // Required methods
    fn read<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_read<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn skip<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_skip<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_to_input_stream<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream
    ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_input_stream<'life0, 'async_trait>(
        &'life0 mut self,
        this: InputStream
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn check_write<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        contents: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_write_and_flush<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        contents: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn flush<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_flush<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_to_output_stream<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream
    ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_zeroes<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn splice<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        src: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_splice<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        src: InputStream,
        len: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn forward<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream,
        src: InputStream
    ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_output_stream<'life0, 'async_trait>(
        &'life0 mut self,
        this: OutputStream
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Required Methods

fn read<'life0, 'async_trait>( &'life0 mut self, this: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Perform a non-blocking read from the stream.

This function returns a list of bytes containing the data that was read, along with a stream-status which, indicates whether further reads are expected to produce data. The returned list will contain up to len bytes; it may return fewer than requested, but not more. An empty list and stream-status:open indicates no more data is available at this time, and that the pollable given by subscribe-to-input-stream will be ready when more data is available.

Once a stream has reached the end, subsequent calls to read or skip will always report stream-status:ended rather than producing more data.

When the caller gives a len of 0, it represents a request to read 0 bytes. This read should always succeed and return an empty list and the current stream-status.

The len parameter is a u64, which could represent a list of u8 which is not possible to allocate in wasm32, or not desirable to allocate as as a return value by the callee. The callee may return a list of bytes less than len in size while more bytes are available for reading.

fn blocking_read<'life0, 'async_trait>( &'life0 mut self, this: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Read bytes from a stream, after blocking until at least one byte can be read. Except for blocking, identical to read.

fn skip<'life0, 'async_trait>( &'life0 mut self, this: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Skip bytes from a stream.

This is similar to the read function, but avoids copying the bytes into the instance.

Once a stream has reached the end, subsequent calls to read or skip will always report end-of-stream rather than producing more data.

This function returns the number of bytes skipped, along with a stream-status indicating whether the end of the stream was reached. The returned value will be at most len; it may be less.

fn blocking_skip<'life0, 'async_trait>( &'life0 mut self, this: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Skip bytes from a stream, after blocking until at least one byte can be skipped. Except for blocking behavior, identical to skip.

fn subscribe_to_input_stream<'life0, 'async_trait>( &'life0 mut self, this: InputStream ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once either the specified stream has bytes available to read or the other end of the stream has been closed. The created pollable is a child resource of the input-stream. Implementations may trap if the input-stream is dropped before all derived pollables created with this function are dropped.

fn drop_input_stream<'life0, 'async_trait>( &'life0 mut self, this: InputStream ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Dispose of the specified input-stream, after which it may no longer be used. Implementations may trap if this input-stream is dropped while child pollable resources are still alive. After this input-stream is dropped, implementations may report any corresponding output-stream has stream-state.closed.

fn check_write<'life0, 'async_trait>( &'life0 mut self, this: OutputStream ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Check readiness for writing. This function never blocks.

Returns the number of bytes permitted for the next call to write, or an error. Calling write with more bytes than this function has permitted will trap.

When this function returns 0 bytes, the subscribe-to-output-stream pollable will become ready when this function will report at least 1 byte, or an error.

fn write<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, contents: Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Perform a write. This function never blocks.

Precondition: check-write gave permit of Ok(n) and contents has a length of less than or equal to n. Otherwise, this function will trap.

returns Err(closed) without writing if the stream has closed since the last call to check-write provided a permit.

fn blocking_write_and_flush<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, contents: Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Perform a write of up to 4096 bytes, and then flush the stream. Block until all of these operations are complete, or an error occurs.

This is a convenience wrapper around the use of check-write, subscribe-to-output-stream, write, and flush, and is implemented with the following pseudo-code:

let pollable = subscribe-to-output-stream(this);
while !contents.is_empty() {
// Wait for the stream to become writable
poll-oneoff(pollable);
let Ok(n) = check-write(this); // eliding error handling
let len = min(n, contents.len());
let (chunk, rest) = contents.split_at(len);
write(this, chunk);            // eliding error handling
contents = rest;
}
flush(this);
// Wait for completion of `flush`
poll-oneoff(pollable);
// Check for any errors that arose during `flush`
let _ = check-write(this);       // eliding error handling

fn flush<'life0, 'async_trait>( &'life0 mut self, this: OutputStream ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Request to flush buffered output. This function never blocks.

This tells the output-stream that the caller intends any buffered output to be flushed. the output which is expected to be flushed is all that has been passed to write prior to this call.

Upon calling this function, the output-stream will not accept any writes (check-write will return ok(0)) until the flush has completed. The subscribe-to-output-stream pollable will become ready when the flush has completed and the stream can accept more writes.

fn blocking_flush<'life0, 'async_trait>( &'life0 mut self, this: OutputStream ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Request to flush buffered output, and block until flush completes and stream is ready for writing again.

fn subscribe_to_output_stream<'life0, 'async_trait>( &'life0 mut self, this: OutputStream ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once the output-stream is ready for more writing, or an error has occured. When this pollable is ready, check-write will return ok(n) with n>0, or an error.

If the stream is closed, this pollable is always ready immediately.

The created pollable is a child resource of the output-stream. Implementations may trap if the output-stream is dropped before all derived pollables created with this function are dropped.

fn write_zeroes<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Write zeroes to a stream.

this should be used precisely like write with the exact same preconditions (must use check-write first), but instead of passing a list of bytes, you simply pass the number of zero-bytes that should be written.

fn splice<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, src: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Read from one stream and write to another.

This function returns the number of bytes transferred; it may be less than len.

Unlike other I/O functions, this function blocks until all the data read from the input stream has been written to the output stream.

fn blocking_splice<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, src: InputStream, len: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Read from one stream and write to another, with blocking.

This is similar to splice, except that it blocks until at least one byte can be read.

fn forward<'life0, 'async_trait>( &'life0 mut self, this: OutputStream, src: InputStream ) -> Pin<Box<dyn Future<Output = Result<Result<(u64, StreamStatus), ()>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Forward the entire contents of an input stream to an output stream.

This function repeatedly reads from the input stream and writes the data to the output stream, until the end of the input stream is reached, or an error is encountered.

Unlike other I/O functions, this function blocks until the end of the input stream is seen and all the data has been written to the output stream.

This function returns the number of bytes transferred, and the status of the output stream.

fn drop_output_stream<'life0, 'async_trait>( &'life0 mut self, this: OutputStream ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Dispose of the specified output-stream, after which it may no longer be used. Implementations may trap if this output-stream is dropped while child pollable resources are still alive. After this output-stream is dropped, implementations may report any corresponding input-stream has stream-state.closed.