Module wasmtime_wasi::preview2::bindings::sync_io::filesystem::types

Re-exports

pub use super::super::super::super::super::filesystem::types::Descriptor;

pub use super::super::super::super::super::filesystem::types::DirectoryEntryStream;

Structs

DescriptorFlags

DescriptorStat

File attributes.

DirectoryEntry

A directory entry.

MetadataHashValue

A 128-bit hash value, split into parts because wasm doesn’t have a 128-bit integer type.

OpenFlags

PathFlags

Enums

Advice

File or memory access pattern advisory information.

DescriptorType

The type of a filesystem object referenced by a descriptor.

ErrorCode

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

NewTimestamp

When setting a timestamp, this gives the value to set it to.

Traits

Host

HostDescriptor

HostDirectoryEntryStream

Functions

add_to_linker

Type Aliases

Datetime

Error

Filesize

File size or length of a region within a file.

InputStream

LinkCount

Number of hard links to an inode.

OutputStream