Trait wasmtime_wasi::preview2::bindings::wasi::filesystem::types::Host

pub trait Host: HostDescriptor + HostDirectoryEntryStream {
    // Required methods
    fn filesystem_error_code(
        &mut self,
        err: Resource<Error>
    ) -> Result<Option<ErrorCode>>;
    fn convert_error_code(&mut self, err: FsError) -> Result<ErrorCode>;
}

Required Methods

fn filesystem_error_code( &mut self, err: Resource<Error> ) -> Result<Option<ErrorCode>>

Attempts to extract a filesystem-related error-code from the stream error provided.

Stream operations which return stream-error::last-operation-failed have a payload with more information about the operation that failed. This payload can be passed through to this function to see if there’s filesystem-related information about the error to return.

Note that this function is fallible because not all stream-related errors are filesystem-related errors.

fn convert_error_code(&mut self, err: FsError) -> Result<ErrorCode>