Module wasmtime_wasi::preview2::pipe

Virtual pipes.

These types provide easy implementations of WasiFile that mimic much of the behavior of Unix pipes. These are particularly helpful for redirecting WASI stdio handles to destinations other than OS files.

Some convenience constructors are included for common backing types like Vec<u8> and String, but the virtual pipes can be instantiated with any Read or Write type.

Structs

AsyncReadStream

Provides a HostInputStream impl from a [tokio::io::AsyncRead] impl

AsyncWriteStream

Provides a HostOutputStream impl from a [tokio::io::AsyncWrite] impl

ClosedInputStream

A stream that is ready immediately, but will always report that it’s closed.

ClosedOutputStream

An output stream that is always closed.

MemoryInputPipe

MemoryOutputPipe

SinkOutputStream

An output stream that consumes all input written to it, and is always ready.

Functions

pipe

TODO