Struct wasmtime_wasi::runtime::AbortOnDropJoinHandle

pub struct AbortOnDropJoinHandle<T>(/* private fields */);

Exactly like a [tokio::task::JoinHandle], except that it aborts the task when the handle is dropped.

This behavior makes it easier to tie a worker task to the lifetime of a Resource by keeping this handle owned by the Resource.