Struct wasmtime_wasi::TrappableError

pub struct TrappableError<T> { /* private fields */ }

A helper error type used by many other modules through type aliases.

This type is an Error itself and is intended to be a representation of either:

A custom error type T

A trap, represented as anyhow::Error

This error is created through either the ::trap constructor representing a full-fledged trap or the From<T> constructor which is intended to be used with ?. The goal is to make normal errors T โautomaticโ but enable error paths to return a ::trap error optionally still as necessary without extra boilerplate everywhere else.

Note that this type isnโt used directly but instead is intended to be used as:

type MyError = TrappableError<bindgen::TheError>;

where MyError is what youโll use throughout bindings code and bindgen::TheError is the type that this represents as generated by the bindgen! macro.

Implementations

impl<T> TrappableError<T>

pub fn trap(err: impl Into<Error>) -> TrappableError<T>

pub fn downcast(self) -> Result<T>
where T: Error + Send + Sync + 'static,

pub fn downcast_ref(&self) -> Option<&T>
where T: Error + Send + Sync + 'static,