Module wasmtime_wasi::sync

Re-export the commonly used wasi-cap-std-sync crate here. This saves consumers of this library from having to keep additional dependencies in sync.

Modules

clocks

dir

file

sched

snapshots

stdio

Structs

Dir
A reference to an open directory on a filesystem.

WasiCtxBuilder

Functions

add_to_linker

clocks_ctx

random_ctx

sched_ctx