Crate wasmtime_wasi

wasmtime-wasi now supports using multiple snapshots to interface to the same WasiCtx!

wasmtime_wasi::Wasi::new(&Store, WasiCtx) is a struct which owns your WasiCtx and provides linkage to every available snapshot.

Individual snapshots are available through wasmtime_wasi::snapshots::preview_{0, 1}::Wasi::new(&Store, Rc<RefCell<WasiCtx>>).

Re-exports

pub use sync::*;

Modules

sync

Re-export the commonly used wasi-cap-std-sync crate here. This saves consumers of this library from having to keep additional dependencies in sync.

Structs

Error

I32Exit

An error returned from the proc_exit host syscall.

WasiCtx

Traits

WasiDir

WasiFile