Struct wasmtime_wasi::old::snapshot_0::Wasi

pub struct Wasi {
    pub args_get: Func,
    pub args_sizes_get: Func,
    pub environ_get: Func,
    pub environ_sizes_get: Func,
    pub clock_res_get: Func,
    pub clock_time_get: Func,
    pub fd_advise: Func,
    pub fd_allocate: Func,
    pub fd_close: Func,
    pub fd_datasync: Func,
    pub fd_fdstat_get: Func,
    pub fd_fdstat_set_flags: Func,
    pub fd_fdstat_set_rights: Func,
    pub fd_filestat_get: Func,
    pub fd_filestat_set_size: Func,
    pub fd_filestat_set_times: Func,
    pub fd_pread: Func,
    pub fd_prestat_get: Func,
    pub fd_prestat_dir_name: Func,
    pub fd_pwrite: Func,
    pub fd_read: Func,
    pub fd_readdir: Func,
    pub fd_renumber: Func,
    pub fd_seek: Func,
    pub fd_sync: Func,
    pub fd_tell: Func,
    pub fd_write: Func,
    pub path_create_directory: Func,
    pub path_filestat_get: Func,
    pub path_filestat_set_times: Func,
    pub path_link: Func,
    pub path_open: Func,
    pub path_readlink: Func,
    pub path_remove_directory: Func,
    pub path_rename: Func,
    pub path_symlink: Func,
    pub path_unlink_file: Func,
    pub poll_oneoff: Func,
    pub proc_exit: Func,
    pub proc_raise: Func,
    pub sched_yield: Func,
    pub random_get: Func,
    pub sock_recv: Func,
    pub sock_send: Func,
    pub sock_shutdown: Func,
}

An instantiated instance of the wasi exports.

This represents a wasi module which can be used to instantiate other wasm modules. This structure exports all that various fields of the wasi instance as fields which can be used to implement your own instantiation logic, if necessary. Additionally Wasi::get_export can be used to do name-based resolution.

Fields
args_get: Funcargs_sizes_get: Funcenviron_get: Funcenviron_sizes_get: Funcclock_res_get: Funcclock_time_get: Funcfd_advise: Funcfd_allocate: Funcfd_close: Funcfd_datasync: Funcfd_fdstat_get: Funcfd_fdstat_set_flags: Funcfd_fdstat_set_rights: Funcfd_filestat_get: Funcfd_filestat_set_size: Funcfd_filestat_set_times: Funcfd_pread: Funcfd_prestat_get: Funcfd_prestat_dir_name: Funcfd_pwrite: Funcfd_read: Funcfd_readdir: Funcfd_renumber: Funcfd_seek: Funcfd_sync: Funcfd_tell: Funcfd_write: Funcpath_create_directory: Funcpath_filestat_get: Funcpath_filestat_set_times: Funcpath_link: Funcpath_open: Funcpath_readlink: Funcpath_remove_directory: Funcpath_rename: Funcpath_symlink: Funcpath_unlink_file: Funcpoll_oneoff: Funcproc_exit: Funcproc_raise: Funcsched_yield: Funcrandom_get: Funcsock_recv: Funcsock_send: Funcsock_shutdown: Func
Implementations

impl Wasi

pub fn new(store: &Store, cx: WasiCtx) -> Wasi

Creates a new Wasi instance.

External values are allocated into the store provided and configuration of the wasi instance itself should be all contained in the cx parameter.

pub fn get_export(&self, name: &str) -> Option<&Func>

Looks up a field called name in this structure, returning it if found.

This is often useful when instantiating a wasmtime instance where name resolution often happens with strings.

pub fn add_to_linker(&self, linker: &mut Linker) -> Result<()>

Adds all wasi items to the specified Linker.