Module wasmtime_wasi::preview2

Wasmtime’s WASI Preview 2 Implementation

Welcome to the (new!) WASI implementation from the Wasmtime team. The goal of this implementation is to support WASI Preview 2 via the Component Model, as well as to provide legacy Preview 1 host support with an adapter that is implemented in terms of the Preview 2 interfaces.

Presently, this crate is experimental. We don’t yet recommend you use it in production. Specifically:

it does not yet support a synchronous rust embedding

polling and streams need a redesign. IO that currently should be non-blocking may be blocking. poll probably doesn’t work at all.

its internal organization could use some love

the wit files in tree describing preview 2 are not faithful to the standards repos

Once these issues are resolved, we expect to move this namespace up to the root of the wasmtime-wasi crate, and move its other exports underneath a pub mod legacy with an off-by-default feature flag, and after 2 releases, retire and remove that code from our tree.

Re-exports

pub use clocks::WasiClocks;

pub use clocks::WasiMonotonicClock;

pub use clocks::WasiWallClock;

pub use stream::InputStream;

pub use stream::OutputStream;

pub use table::Table;

pub use table::TableError;

Modules

clocks

pipe

Virtual pipes.

preview1

preview2

random

stdio

stream

table

wasi

Structs

DirPerms

FilePerms

I32Exit

An error returned from the proc_exit host syscall.

WasiCtx

WasiCtxBuilder

Enums

SystemTimeSpec

A value for specifying a time.

Traits

RngCore

The core of a random number generator.

WasiView