Module wasmtime_wasi::preview2

Wasmtime’s WASI Preview 2 Implementation

Welcome to the (new!) WASI implementation from the Wasmtime team. The goal of this implementation is to support WASI Preview 2 via the Component Model, as well as to provide legacy Preview 1 host support with an adapter that is implemented in terms of the Preview 2 interfaces.

Presently, this crate is experimental. We don’t yet recommend you use it in production. Specifically:

the wit files in tree describing preview 2 are not faithful to the standards repos

Once these issues are resolved, we expect to move this namespace up to the root of the wasmtime-wasi crate, and move its other exports underneath a pub mod legacy with an off-by-default feature flag, and after 2 releases, retire and remove that code from our tree.

Modules

bindings

command

pipe

Virtual pipes.

preview1

Structs

Deterministic

Implement insecure-random using a deterministic cycle of bytes.

DirPerms

FilePerms

I32Exit

An error returned from the proc_exit host syscall.

OccupiedEntry

Like std::collections::hash_map::OccupiedEntry, with a subset of methods available in order to uphold Table invariants.

Table

The Table type is designed to map u32 handles to resources. The table is now part of the public interface to a WasiCtx - it is reference counted so that it can be shared beyond a WasiCtx with other WASI proposals (e.g. wasi-crypto and wasi-nn) to manage their resources. Elements in the Table are Any typed.

WasiCtx

WasiCtxBuilder

Enums

HostPollable

A host representation of the wasi:poll/poll.pollable resource.

StreamState

SystemTimeSpec

A value for specifying a time.

TableError

Traits

HostInputStream

Host trait for implementing the wasi:io/streams.input-stream resource: A bytestream which can be read from.

HostMonotonicClock

HostOutputStream

Host trait for implementing the wasi:io/streams.output-stream resource: A bytestream which can be written to.

HostWallClock

RngCore

The core of a random number generator.

TablePollableExt

TableStreamExt

Extension trait for managing HostInputStreams and HostOutputStreams in the Table.

WasiView

Functions

thread_rng

Type Definitions

ClosureFuture

MakeFuture

PollableFuture