Module wasmtime_wasi::preview2::bindings::sync_io::filesystem::types

Structs

DescriptorFlags

DescriptorStat

File attributes.

DirectoryEntry

A directory entry.

Error

MetadataHashValue

A 128-bit hash value, split into parts because wasm doesn’t have a 128-bit integer type.

Modes

OpenFlags

PathFlags

Enums

AccessType

Access type used by access-at.

Advice

File or memory access pattern advisory information.

DescriptorType

The type of a filesystem object referenced by a descriptor.

ErrorCode

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

NewTimestamp

When setting a timestamp, this gives the value to set it to.

Traits

Host

Functions

add_to_linker

Type Definitions

Datetime

Descriptor

A descriptor is a reference to a filesystem object, which may be a file, directory, named pipe, special file, or other object on which filesystem calls may be made.

DirectoryEntryStream

A stream of directory entries.

Filesize

File size or length of a region within a file.

InputStream

LinkCount

Number of hard links to an inode.

OutputStream