Struct wasmtime_wasi::WasiCtx

pub struct WasiCtx(_);

An Arc-wrapper around the wasi-common context to allow mutable access to the file descriptor table. This wrapper is only necessary due to the signature of fd_fdstat_set_flags; if that changes, there are a variety of improvements that can be made (TODO: https://github.com/bytecodealliance/wasmtime/issues/5643).

Implementations

impl WasiCtx

pub fn new( random: Box<dyn RngCore + Sync + Send, Global>, clocks: WasiClocks, sched: Box<dyn WasiSched, Global>, table: Table ) -> WasiCtx

pub fn insert_file( &self, fd: u32, file: Box<dyn WasiFile, Global>, access_mode: FileAccessMode )

pub fn push_file( &self, file: Box<dyn WasiFile, Global>, access_mode: FileAccessMode ) -> Result<u32, Error>

pub fn insert_dir(&self, fd: u32, dir: Box<dyn WasiDir, Global>, path: PathBuf)

pub fn push_dir( &self, dir: Box<dyn WasiDir, Global>, path: PathBuf ) -> Result<u32, Error>

pub fn table(&self) -> &Table

pub fn table_mut(&mut self) -> Option<&mut Table>

pub fn push_arg(&mut self, arg: &str) -> Result<(), StringArrayError>

pub fn push_env( &mut self, var: &str, value: &str ) -> Result<(), StringArrayError>

pub fn set_stdin(&self, f: Box<dyn WasiFile, Global>)

pub fn set_stdout(&self, f: Box<dyn WasiFile, Global>)

pub fn set_stderr(&self, f: Box<dyn WasiFile, Global>)

pub fn push_preopened_dir( &self, dir: Box<dyn WasiDir, Global>, path: impl AsRef<Path> ) -> Result<(), Error>