Module wasmtime_wasi::sync

Re-export the commonly used wasi-cap-std-sync crate here. This saves consumers of this library from having to keep additional dependencies in sync.

Modules

clocks

dir

file

net

sched

snapshots

stdio

Structs

Dir

A reference to an open directory on a filesystem.

TcpListener

A TCP socket server, listening for connections.

WasiCtxBuilder

Functions

add_to_linker

ambient_authority

Return an AmbientAuthority value, which allows use of functions that include an AmbientAuthority argument.

clocks_ctx

random_ctx

sched_ctx